+++ quantiphy/__init__.py
@@ -1,4 +1,5 @@
 from .quantiphy import (
+
     QuantiPhyError,      # exceptions
     ExpectedQuantity,
     IncompatibleUnits,
@@ -11,15 +12,20 @@ from .quantiphy import (
     UnknownScaleFactor,
     UnknownUnitSystem,
     IncompatiblePreferences,
+
     UnitConversion,      # unit conversions
     set_unit_system,
+
     add_constant,        # constants
+
     Quantity,            # quantities
+
     as_real,             # quantity functions
     as_tuple,
     render,
     fixed,
     binary,
+
     __version__,         # version
     __released__,
 )