+++ tests/test_unit_conversion.py
@@ -369,35 +369,34 @@ def test_affine_conversion():
     result = conversion.convert(32, from_units='F')
     assert str(result) == '0 C'
 
-    with pytest.raises(UnknownConversion) as exception:
-        result = conversion.convert(0, from_units='X', to_units='X')
-    assert str(exception.value) == "unable to convert to ‘X’."
+    result = conversion.convert(0, from_units='X', to_units='X')
+    assert str(result) == '0 X'
 
     with pytest.raises(UnknownConversion) as exception:
         result = conversion.convert(0, from_units='F', to_units='X')
-    assert str(exception.value) == "unable to convert to ‘X’."
+    assert str(exception.value) == "unable to convert between ‘X’ and ‘F’."
 
     with pytest.raises(UnknownConversion) as exception:
         result = conversion.convert(0, from_units='X', to_units='F')
-    assert str(exception.value) == "unable to convert from ‘X’."
+    assert str(exception.value) == "unable to convert between ‘F’ and ‘X’."
     assert isinstance(exception.value, UnknownConversion)
     assert isinstance(exception.value, QuantiPhyError)
     assert isinstance(exception.value, KeyError)
     assert exception.value.args == ()
-    assert exception.value.kwargs == dict(from_units='X',)
+    assert exception.value.kwargs == dict(from_units='X', to_units='F')
 
     with pytest.raises(UnknownConversion) as exception:
         result = conversion.convert(0, to_units='X')
-    assert str(exception.value) == "unable to convert to ‘X’."
+    assert str(exception.value) == "unable to convert between ‘X’ and ‘F’."
 
     with pytest.raises(KeyError) as exception:
         result = conversion.convert(0, from_units='X')
-    assert str(exception.value) == "unable to convert from ‘X’."
+    assert str(exception.value) == "unable to convert between ‘C’ and ‘X’."
     assert isinstance(exception.value, UnknownConversion)
     assert isinstance(exception.value, QuantiPhyError)
     assert isinstance(exception.value, KeyError)
     assert exception.value.args == ()
-    assert exception.value.kwargs == dict(from_units='X',)
+    assert exception.value.kwargs == dict(from_units='X', to_units='C')
 
 def test_func_converters():
     Quantity.reset_prefs()